requirements.txt
# YOLO Object Detection System Requirements

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# YOLO Models
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Development Tools (Optional)
pytest>=7.4.0
black>=23.0.0
mypy>=1.5.0
